% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus-zoom.R
\name{create_gene_figure}
\alias{create_gene_figure}
\title{Display genes based on position in the genome}
\usage{
create_gene_figure(
  region_chr,
  region_start,
  region_end,
  plot_start_limit,
  plot_end_limit,
  build,
  add_region,
  gene_type_in_name = TRUE,
  ...
)
}
\arguments{
\item{region_chr}{Chromosome, UCSC format: "chr11" not 11}

\item{region_start}{start of region}

\item{region_end}{end of region}

\item{plot_start_limit}{start of figure, in genomic location}

\item{plot_end_limit}{end of figure, in genomic location}

\item{build}{GRCh38 or GRCh37 to map genes to regions?}

\item{add_region}{input a tibble in the same format to add a custom region to draw}

\item{gene_type_in_name}{append the gene_type to the y-axis label on the gene plot}

\item{...}{pass additional arguments to get_genes_in_region, for example gene_type = "protein_coding"}
}
\value{
a ggplot2 object
}
\description{
Display genes based on position in the genome
}
\examples{
\dontrun{
create_gene_figure(chr = "chr12", 10, 100000, build = 37)
}
}
