% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus-zoom.R
\name{locus_zoom}
\alias{locus_zoom}
\title{Plot a GWAS loci colored by ld with nearby SNPs}
\usage{
locus_zoom(gwas, bed_file, build = c("38", "37"), expand = 100000L, ld_path)
}
\arguments{
\item{gwas}{a summary statistics file in tidyGWAS format}

\item{bed_file}{a bed file with six columns}

\item{build}{genome build of GWAS and bedfile}

\item{expand}{how much to expand sides of loci}

\item{ld_path}{filepath to LD reference data}
}
\value{
a [[ggplot2::ggplot()]] figure
}
\description{
Plot a GWAS loci colored by ld with nearby SNPs
}
\examples{
\dontrun{
locus_zoom(gwas, bedfile, build = "38", ld_path ="path/to/topmed_ld")
}
}
