% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld-query.R
\name{query_ld}
\alias{query_ld}
\title{Get the LD for a index snp, defined by chr and pos, using reference LD data}
\usage{
query_ld(chr, pos, ld_ref)
}
\arguments{
\item{chr}{chromosome in integer format (No support yet for X and Y)}

\item{pos}{position}

\item{ld_ref}{filepath to the ld reference data}
}
\value{
a [dplyr::tibble()]
}
\description{
Get the LD for a index snp, defined by chr and pos, using reference LD data
}
\examples{
\dontrun{
query_ld(chr = 7, pos = 24398341, ld_ref = "ld_ref_topmed/ld_single_nucleotide_variants/")
}
}
