% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bed.R
\name{read_bed}
\alias{read_bed}
\title{Read a bed file into a tibble}
\usage{
read_bed(path, ...)
}
\arguments{
\item{path}{filepath}

\item{...}{arguments to pass to [arrow::read_tsv_arrow()]}
}
\value{
a tibble
}
\description{
Read a bed file into a tibble
}
\examples{
\dontrun{
read_bed("file.bed")
}
}
